Fixed Kotlin Master and switched back Npm -> Yarn (resolved KT-78504)
