+++ compiler-plugin/compiler-plugin-cli/src/main/kotlin/kotlinx/rpc/codegen/RpcCommandLineProcessor.kt
@@ -4,18 +4,14 @@
 
 package kotlinx.rpc.codegen
 
-import kotlinx.rpc.codegen.extension.RpcIrExtension
-import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.compiler.plugin.AbstractCliOption
 import org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
-import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
 import org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi
 import org.jetbrains.kotlin.config.CompilerConfiguration
-import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrarAdapter
 
 @OptIn(ExperimentalCompilerApi::class)
 class RpcCommandLineProcessor : CommandLineProcessor {
-    override val pluginId = "kotlinx-rpc"
+    override val pluginId = PLUGIN_ID
 
     override val pluginOptions = listOf(
         RpcFirCliOptions.ANNOTATION_TYPE_SAFETY,
@@ -36,21 +32,4 @@ class RpcCommandLineProcessor : CommandLineProcessor {
             }
         }
     }
-}
-
-@OptIn(ExperimentalCompilerApi::class)
-class RpcCompilerPlugin : CompilerPluginRegistrar() {
-    override val supportsK2: Boolean = true
-
-    override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
-        registerRpcExtensions(configuration)
-    }
-}
-
-@OptIn(ExperimentalCompilerApi::class)
-fun CompilerPluginRegistrar.ExtensionStorage.registerRpcExtensions(configuration: CompilerConfiguration) {
-    VersionSpecificApi.INSTANCE = VersionSpecificApiImpl
-
-    IrGenerationExtension.registerExtension(RpcIrExtension(configuration))
-    FirExtensionRegistrarAdapter.registerExtension(FirRpcExtensionRegistrar(configuration))
 }
