+++ compiler-plugin/compiler-plugin-cli/src/main/kotlin/kotlinx/rpc/codegen/pluginId.kt
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2023-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package kotlinx.rpc.codegen
+
+const val PLUGIN_ID = "org.jetbrains.kotlinx.rpc"
