+++ compiler-plugin/compiler-plugin-k2/src/main/kotlin/kotlinx/rpc/codegen/FirRpcServiceGenerator.kt
@@ -96,7 +96,7 @@ class FirRpcServiceGenerator(
      * Generates [owner]'s service stub.
      * Scrapes the functions from the [owner] to generate method classes.
      */
-    private fun generateRpcServiceStubClass(owner: FirClassSymbol<*>): FirRegularClassSymbol? {
+    private fun generateRpcServiceStubClass(owner: FirClassSymbol<*>): FirRegularClassSymbol {
         return createNestedClass(owner, RpcNames.SERVICE_STUB_NAME, RpcGeneratedStubKey(owner.name)) {
             visibility = Visibilities.Public
             modality = Modality.FINAL
