+++ gradle.properties
@@ -20,9 +20,6 @@ org.gradle.configuration-cache=true
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
 
-# https://youtrack.jetbrains.com/issue/KT-78504
-kotlin.js.yarn=false
-
 # development mode for kotlinx.rpc gradle plugin. Uses local project paths to apply the compiler plugin
 kotlinx.rpc.plugin.internalDevelopment=true
